+++ R/isoweek-class.R
@@ -23,7 +23,7 @@
 #'
 #' `isoweek()` is a constructor for `<grates_isoweek>` objects. It takes a
 #' vector of year and vector of week values as inputs. Length 1 inputs will be
-#' recycled to the length of the other input and `double` vectors will again be
+#' recycled to the length of the other input and `double` vectors will be
 #' converted to integer via `as.integer(floor(x))`.
 #'
 #' `as_isoweek()` is a generic for conversion to `<grates_isoweek>`.
@@ -83,7 +83,8 @@
 #' @references
 #' Wikipedia contributors. (2025, January 15). ISO week date.
 #' In Wikipedia, The Free Encyclopedia.
-#' Retrieved 09:19, March 6, 2025, from https://en.wikipedia.org/w/index.php?title=ISO_week_date&oldid=1269568343
+#' Retrieved 09:19, March 6, 2025, from
+#' https://en.wikipedia.org/w/index.php?title=ISO_week_date&oldid=1269568343
 #'
 # -------------------------------------------------------------------------
 #' @examples
@@ -97,7 +98,7 @@
 #' # character coercion assumes date input by default
 #' as_isoweek("2019-05-03")
 #'
-#' # character coercion can handle YYYY-Www format too
+#' # character coercion can handle YYYY-Wxx format too
 #' as_isoweek("2019-W12", format = "yearweek")
 #'
 #' # construction
@@ -120,19 +121,8 @@ NULL
 #' @export
 isoweek <- function(year = integer(), week = integer()) {
 
-    # check year is integerish
-    if (is.vector(year, "double")) {
-        year <- as.integer(floor(year))
-    } else if (!is.integer(year)) {
-        stop("`year` must be integer.")
-    }
-
-    # check week is integerish
-    if (is.vector(week, "double")) {
-        week <- as.integer(floor(week))
-    } else if (!is.integer(week)) {
-        stop("`week` must be integer.")
-    }
+    year <- .make_floored_integer(year)
+    week <- .make_floored_integer(week)
 
     # check compatible lengths
     tmp <- .recycle(year, week)
@@ -217,11 +207,7 @@ as_isoweek.factor <- function(
 #' @rdname isoweek_class
 #' @export
 new_isoweek <- function(x = integer()) {
-    if (is.vector(x, "double")) {
-        x <- as.integer(floor(x))
-    } else if (!is.integer(x)) {
-        stop("`x` must be integer.")
-    }
+    x <- .make_floored_integer(x)
     .new_isoweek(x)
 }
 
@@ -258,27 +244,23 @@ vec_ptype_full.grates_isoweek <- function(x, ...) {"isoweek"}
 #' @export
 `[.grates_isoweek` <- function(x, ..., drop = FALSE) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
 #' @export
 `[[.grates_isoweek` <- function(x, ..., drop = TRUE) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
 #' @export
 `[<-.grates_isoweek` <- function(x, ..., value) {
-    old_class <- class(x)
     if (!inherits(value, "grates_isoweek"))
-        stop("Can only assign <grates_isoweek> objects in to an <grates_isoweek> object.")
+        stop("Can only assign a <grates_isoweek> object into a <grates_isoweek> object.")
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
@@ -289,16 +271,14 @@ vec_ptype_full.grates_isoweek <- function(x, ...) {"isoweek"}
 #' @export
 rep.grates_isoweek <- function(x, ...) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
 #' @export
 unique.grates_isoweek <- function(x, incomparables = FALSE, ...) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
@@ -390,24 +370,21 @@ as.data.frame.grates_isoweek <- as.data.frame.vector
 #' @export
 min.grates_isoweek <- function(x, ..., na.rm = FALSE) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
 #' @export
 max.grates_isoweek <- function(x, ..., na.rm = FALSE) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
 #' @export
 range.grates_isoweek <- function(x, ..., na.rm = FALSE) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
@@ -471,7 +448,7 @@ Ops.grates_isoweek <- function(e1, e2) {
             } else if (inherits(e1, "grates_isoweek") && .is_whole(e2)) {
                 return(.new_isoweek(unclass(e1) - as.integer(e2)))
             }
-            stop("Can only subtract whole numbers and other <grates_isoweek> objects from <grates_isoweek> objects.")
+            stop("Can only subtract whole numbers and other <grates_isoweek> objects from <grates_isoweek> objects.") # nolint: line_length_linter.
         },
         stopf("%s is not compatible with <grates_isoweek> objects.", op)
     )
@@ -489,13 +466,9 @@ is.numeric.grates_isoweek <- function(x) {
 # -------------------------------- INTERNALS ------------------------------ #
 # ------------------------------------------------------------------------- #
 # ------------------------------------------------------------------------- #
-.new_isoweek <- function(x) {
-    class(x) <- "grates_isoweek"
-    x
-}
+.new_isoweek <- function(x) {`class<-`(x, "grates_isoweek")}
 
 .isoweek <- function(year, week) {
     out <- .yearweek(year = year, week = week, firstday = 1L)
-    class(out) <- "grates_isoweek"
-    out
+    `class<-`(out, "grates_isoweek")
 }