+++ R/year-class.R
@@ -12,7 +12,7 @@
 #' `year()` takes as input a vector representing, unsurprisingly, the years.
 #' `double` vectors are coerced via `as.integer(floor(x))`.
 #'
-#' `as_yearquarter()` is a generic for coercing input in to `<grates_year>`.
+#' `as_yearquarter()` is a generic for coercing input into `<grates_year>`.
 #' - Character input is first parsed using `as.Date()`.
 #' - POSIXct and POSIXlt are converted with their timezone respected.
 #'
@@ -53,12 +53,7 @@ NULL
 #' @rdname year_class
 #' @export
 year <- function(x = integer()) {
-    if (is.vector(x, "double")) {
-        x <- as.integer(floor(x))
-    } else if (!is.integer(x)) {
-        stop("`x` must be integer.")
-    }
-
+    x <- .make_floored_integer(x)
     .new_year(x = x)
 }
 
@@ -158,26 +153,23 @@ vec_ptype_full.grates_year <- function(x, ...) {"grates_year"}
 #' @export
 `[.grates_year` <- function(x, ..., drop = FALSE) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
 #' @export
 `[[.grates_year` <- function(x, ..., drop = TRUE) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
 #' @export
 `[<-.grates_year` <- function(x, ..., value) {
     if (!inherits(value, "grates_year"))
-        stop("Can only assign <grates_year> objects in to an <grates_year> object.")
+        stop("Can only assign a <grates_year> object into a <grates_year> object.")
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
@@ -188,16 +180,14 @@ vec_ptype_full.grates_year <- function(x, ...) {"grates_year"}
 #' @export
 rep.grates_year <- function(x, ...) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
 #' @export
 unique.grates_year <- function(x, incomparables = FALSE, ...) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
@@ -251,8 +241,12 @@ as.Date.grates_year <- function(x, ...) {
 # -------------------------------------------------------------------------
 #' @export
 as.POSIXct.grates_year <- function(x, tz = "UTC", ...) {
-    if (tz != "UTC")
-        stop("<grates_year> objects can only be converted to UTC. If other timezones are required, first convert to <Date> and then proceed as desired.")
+    if (tz != "UTC") {
+        stop(
+            "<grates_year> objects can only be converted to UTC. ",
+            "If other timezones are required, first convert to <Date> and then proceed as desired."
+        )
+    }
     x <- .month_to_days((unclass(x) - 1970L) * 12L)
     .POSIXct(x * 86400, tz = "UTC")
 }
@@ -260,8 +254,13 @@ as.POSIXct.grates_year <- function(x, tz = "UTC", ...) {
 # -------------------------------------------------------------------------
 #' @export
 as.POSIXlt.grates_year <- function(x, tz = "UTC", ...) {
-    if (tz != "UTC")
-        stop("<grates_year> objects can only be converted to UTC. If other timezones are required, first convert to <Date> and then proceed as desired.")
+    if (tz != "UTC") {
+        stop(
+            "<grates_year> objects can only be converted to UTC. ",
+            "If other timezones are required, first convert to <Date> and then proceed as desired."
+        )
+    }
+
     x <- .month_to_days((unclass(x) - 1970L) * 12L)
     as.POSIXlt(.POSIXct(x * 86400, tz = "UTC"), tz = "UTC")
 }
@@ -286,24 +285,21 @@ as.data.frame.grates_year <- as.data.frame.vector
 #' @export
 min.grates_year <- function(x, ..., na.rm = FALSE) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
 #' @export
 max.grates_year <- function(x, ..., na.rm = FALSE) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
 #' @export
 range.grates_year <- function(x, ..., na.rm = FALSE) {
     out <- NextMethod()
-    class(out) <- class(x)
-    out
+    `class<-`(out, class(x))
 }
 
 # -------------------------------------------------------------------------
@@ -363,7 +359,7 @@ Ops.grates_year <- function(e1, e2) {
             } else if (inherits(e1, "grates_year") && .is_whole(e2)) {
                 return(.new_year(unclass(e1) - as.integer(e2)))
             }
-            stop("Can only subtract whole numbers and other <grates_year> objects from <grates_year> objects.")
+            stop("Can only subtract whole numbers and other <grates_year> objects from <grates_year> objects.") # nolint: line_length_linter.
         },
         stopf("%s is not compatible with <grates_year> objects.", op)
     )
@@ -381,7 +377,4 @@ is.numeric.grates_year <- function(x) {
 # ------------------------------------------------------------------------- #
 # ------------------------------------------------------------------------- #
 
-.new_year <- function(x = integer()) {
-    class(x) <- "grates_year"
-    x
-}
+.new_year <- function(x = integer()) {`class<-`(x, "grates_year")}