+++ NEWS.md
@@ -1,3 +1,15 @@
+# grates 1.6.0
+
+## breaking change
+
+* The default `offset` argument for the Date and POSIXt methods of the
+  `as_period()` generic is now set as the minimum of the input argument.
+  Motivated by discussion with @joshwlambert.
+  
+* For `as_period()` methods, `offset` can now be passed as a character or
+  factor value that is coercible to a Date. Again motivated by discussion with
+  @joshwlambert.
+
 # grates 1.5.0
 
 ## breaking change